+++ include/rocket_fiber.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void rocket_fiber_yield();
