+++ include_internal/internal/pal.h
